src/lw_aifw: Public classes of AI Framework moved to include/lw_aifw
framework/src/lw_aifw:
Public classes of AIFramework moved to framework/include/lw_aifw

apps/examples/lw_aifw_sine_test_app:
DataSource moved to application scope
Modify sample application to sync with AI Framework interfaces
