+++ apps/examples/lw_aifw_sine_test_app/Makefile
@@ -63,6 +63,7 @@ ASRCS	=
 CXXSRCS	=
 
 CFLAGS +=
+CXXFLAGS += -Wno-pedantic
 
 MAINSRC = lwaifw_sine_test_main.cpp
 CXXSRCS +=
