+++ framework/Makefile
@@ -29,6 +29,10 @@ VPATH = .
 
 # Framework Directories
 
+CXXFLAGS += -I../external/tfmicro
+CXXFLAGS += -I../external/tfmicro/third_party/flatbuffers/include
+CXXFLAGS += -I../external/tfmicro/third_party/gemmlowp
+
 # BUILDIRS is the list of top-level directories containing Make.defs files
 
 BUILDIRS   := $(dir $(wildcard src/*/Make.defs))
