+++ framework/include/lw_aifw/AICSVDataSource.h
@@ -17,24 +17,25 @@
  ****************************************************************************/
 
 #pragma once
-
-#include <stdint.h>
 #include "lw_aifw/lw_aifw_result.h"
+#include "lw_aifw/AIDataSource.h"
 
 namespace lw_aifw {
-	typedef void (*rawDataCollectedListener)(LW_AIFW_RESULT result, float *data, uint16_t dataCount, void *args);
-	
-	class AICSVReader {
-	protected:
-		rawDataCollectedListener mRawDataCollectedListener;
-		/* ToDo: plan to use std::function & remove mRawDataCollectedCBArgs */
-		void *mRawDataCollectedCBArgs;
-	public:
-		AICSVReader();
-		virtual ~AICSVReader();
-		virtual void setRawDataCollectedListener(rawDataCollectedListener listener, void *calbackFuncArgs);
-		virtual LW_AIFW_RESULT init(void) = 0;
-		virtual LW_AIFW_RESULT getSourceData(void) = 0;
-	};
+
+class AICSVReader;
+class AICSVDataSource : public AIDataSource {
+private:
+	uint16_t m_SensorCount;
+	std::shared_ptr<AICSVReader> mAICSVReader;
+public:
+	AICSVDataSource(const char *sourceName);
+	AICSVDataSource(float *datavalues, uint16_t sensorCount, uint16_t rowCount);
+	~AICSVDataSource();
+	std::shared_ptr<AICSVReader> getCSVReader(void);
+	LW_AIFW_RESULT shareData(void* data);
+	LW_AIFW_RESULT getDataAsync(void);
+	static void onRawDataCollectedListener(LW_AIFW_RESULT result, float *data, uint16_t dataCount, void *args);
+};
 
 } /* lw_aifw */
+
